start shopping validator Method CreateProduct Method toString SRT
