+++ src/com/javaguru/shoppinglist/Product.java
@@ -67,8 +67,6 @@ public class Product {
     }
 
 
-
-
     @Override
     public String toString() {
         return "Product{" +
